requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio_server]
python-engineio[asyncio]
sqlalchemy[asyncio]
asyncpg
pydantic
firebase-admin
google-auth-oauthlib
google-api-python-client
requests
apscheduler
zstandard
python-dotenv
python-multipart
couchbase
pyjwt
httpx